@@ -1,8 +1,10 @@
 <x-filament-tree::actions.action
     :action="$action"
-    component="filament::button"
+    dynamic-component="filament::button"
     :outlined="$isOutlined()"
+    :labeled-from="$getLabeledFromBreakpoint()"
     :icon-position="$getIconPosition()"
+    :icon-size="$getIconSize()"
     class="filament-tree-button-action"
 >
     {{ $getLabel() }}
