@@ -1,6 +1,6 @@
 <x-filament-tree::actions.action
     :action="$action"
     :label="$getLabel()"
-    component="filament::icon-button"
+    dynamic-component="filament::icon-button"
     class="filament-tree-icon-button-action"
 />
