@@ -1,7 +1,8 @@
 <x-filament-tree::actions.action
     :action="$action"
-    component="filament::link"
+    dynamic-component="filament::link"
     :icon-position="$getIconPosition()"
+    :icon-size="$getIconSize()"
     class="filament-tree-link-action"
 >
     {{ $getLabel() }}
