@@ -8,12 +8,12 @@
     'trigger',
 ])
 
-<x-filament-support::modal
+<x-filament::modal
     :attributes="\Filament\Support\prepare_inherited_attributes($attributes)->merge($slots)"
-    :dark-mode="config('filament.dark_mode')"
+    :dark-mode="\Filament\Facades\Filament::hasDarkMode()"
     heading-component="filament-tree::modal.heading"
     {{-- hr-component="tables::hr" --}}
     subheading-component="filament-tree::modal.subheading"
 >
     {{ $slot }}
-</x-filament-support::modal>
+</x-filament::modal>
