@@ -3,10 +3,9 @@
 namespace SolutionForest\FilamentTree\Actions;
 
 use Closure;
-use Filament\Forms\ComponentContainer;
-use Filament\Support\Actions\Concerns\CanCustomizeProcess;
+use Filament\Actions\Concerns\CanCustomizeProcess;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Arr;
+use SolutionForest\FilamentTree\Components\Tree;
 
 class EditAction extends Action
 {
@@ -25,34 +24,39 @@ class EditAction extends Action
     {
         parent::setUp();
 
-        $this->label(__('filament-support::actions/edit.single.label'));
+        $this->label(__('filament-actions::edit.single.label'));
 
-        $this->modalHeading(fn (): string => __('filament-support::actions/edit.single.modal.heading', ['label' => $this->getRecordTitle()]));
+        $this->modalHeading(fn (): string => __('filament-actions::edit.single.modal.heading', ['label' => $this->getRecordTitle()]));
 
-        $this->icon('heroicon-s-pencil');
+        $this->modalSubmitActionLabel(__('filament-actions::edit.single.modal.actions.save.label'));
 
-        $this->iconButton();
+        $this->successNotificationTitle(__('filament-actions::edit.single.notifications.saved.title'));
 
-        $this->successNotificationTitle(__('filament-support::actions/edit.single.messages.saved'));
+        $this->icon('heroicon-m-pencil-square');
 
-        $this->mountUsing(function (ComponentContainer $form, Model $record): void {
-            $data = $record->attributesToArray();
+        $this->fillForm(function (Model $record, Tree $tree): array {
+            if ($translatableContentDriver = $tree->makeFilamentTranslatableContentDriver()) {
+                $data = $translatableContentDriver->getRecordAttributesToArray($record);
+            } else {
+                $data = $record->attributesToArray();
+            }
 
             if ($this->mutateRecordDataUsing) {
                 $data = $this->evaluate($this->mutateRecordDataUsing, ['data' => $data, 'record' => $record]);
             }
 
-            $form->fill($data);
+            return $data;
         });
 
         $this->action(function (): void {
-            $this->process(function (array $data, Model $record) {
-                if ($this->mutateFormDataBeforeSaveUsing) {
-                    $data = $this->evaluate($this->mutateFormDataBeforeSaveUsing, ['data' => $data]);
+            $this->process(function (array $data, Model $record, Tree $tree) {
+                if ($translatableContentDriver = $tree->makeFilamentTranslatableContentDriver()) {
+                    $translatableContentDriver->updateRecord($record, $data);
+                } else {
+                    $record->update($data);
                 }
-                $record->update($data);
             });
-            
+
             $this->success();
         });
     }
