@@ -4,6 +4,8 @@ namespace SolutionForest\FilamentTree\Widgets;
 
 use Filament\Forms\Concerns\InteractsWithForms;
 use Filament\Forms\Contracts\HasForms;
+use Filament\Infolists\Components\Component as InfolistsComponent;
+use Filament\Support\Contracts\TranslatableContentDriver;
 use Filament\Widgets\Widget;
 use Illuminate\Database\Eloquent\Model;
 use SolutionForest\FilamentTree\Actions\Action;
@@ -11,12 +13,12 @@ use SolutionForest\FilamentTree\Actions\DeleteAction;
 use SolutionForest\FilamentTree\Actions\EditAction;
 use SolutionForest\FilamentTree\Actions\ViewAction;
 use SolutionForest\FilamentTree\Components\Tree as TreeComponent;
-use SolutionForest\FilamentTree\Concern;
+use SolutionForest\FilamentTree\Concern\InteractWithTree;
 use SolutionForest\FilamentTree\Contract\HasTree;
 
 class Tree extends Widget implements HasTree, HasForms
 {
-    use Concern\InteractWithTree;
+    use InteractWithTree;
     use InteractsWithForms;
 
     protected static string $view = 'filament-tree::widgets.tree';
@@ -149,7 +151,13 @@ class Tree extends Widget implements HasTree, HasForms
             $schema = $this->getFormSchema();
         }
 
-        $action->form($schema);
+        $action->form($this->getFormSchema());
+
+        $isInfoList = count(array_filter($schema, fn ($component) => $component instanceof InfolistsComponent)) > 0;
+
+        if ($isInfoList) {
+            $action->infolist($schema);
+        }
 
         $action->model($this->getModel());
 
@@ -180,5 +188,10 @@ class Tree extends Widget implements HasTree, HasForms
         }
 
         $this->{$hook}();
+    }
+    
+    public function makeTranslatableContentDriver(): ?TranslatableContentDriver
+    {
+        return null;
     }
 }
